Fixed id mvc for products

+++ server/migrations/20240208123226-create-product.js
@@ -19,10 +19,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       fullPrice: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       screen: {
         type: Sequelize.STRING
